int id -> long id

+++ BestMoviesAPI/main/java/com/bestmovies/sep6_project/controllers/DirectorController.java
@@ -27,17 +27,17 @@ public class DirectorController {
     }
 
     @RequestMapping(value = "/directorId/{directorId}", method = RequestMethod.PUT, consumes="application/json")
-    public void updateMovie(@RequestBody Director updatedDirector, @PathVariable int directorId, HttpServletResponse response){
+    public void updateMovie(@RequestBody Director updatedDirector, @PathVariable long directorId, HttpServletResponse response){
         setResponse(response, directorService.editDirector(updatedDirector ,directorId));
     }
 
     @RequestMapping(value = "/directorId/{directorId}", method = RequestMethod.DELETE)
-    public void deleteDirector(@PathVariable int directorId, HttpServletResponse response){
+    public void deleteDirector(@PathVariable long directorId, HttpServletResponse response){
         setResponse(response, directorService.deleteDirector(directorId));
     }
 
     @RequestMapping(value = "/{id}",method = RequestMethod.GET)
-    Director getDirectorById(@PathVariable int id){
+    Director getDirectorById(@PathVariable long id){
         return directorService.getDirectorById(id);
     }
